# docs/swagger.yaml
basePath: /api/v1        // Base path for all API routes
definitions:
  types.AddPoiRequest:
    properties:
      poi_id:
        type: string
    type: object
...
  types.Response:
    properties:
      error:
        description: Optional error message.
        example: Resource not found
        type: string
      message:
        description: Optional success message.
        example: Operation successful
        type: string
      success:
        description: Indicates if the operation was successful.
        example: true
        type: boolean
    type: object
...
  types.UpdateProfileParams:
    properties:
      about_you:
        type: string
      age:
        type: integer
      city:
        type: string
      country:
        type: string
      displayName:
        type: string
      email:
        type: string
      firstname:
        type: string
      lastname:
        type: string
      profileImageURL:
        type: string
      username:
        description: Pointer allows distinguishing between empty string and not provided
        type: string
    type: object
...
  types.UserProfile:
    properties:
      about_you:
        type: string
      age:
        type: integer
      city:
        type: string
      country:
        type: string
      created_at:
        type: string
      display_name:
        description: Use pointer if nullable
        type: string
      email:
        type: string
      email_verified_at:
        description: Use pointer if nullable
        type: string
      firstname:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      last_login_at:
        description: Use pointer if nullable
        type: string
      lastname:
        type: string
      profile_image_url:
        description: Use pointer if nullable
        type: string
      updated_at:
        type: string
      username:
        description: Use pointer if nullable/optional unique
        type: string
    type: object
...
host: localhost:8000 // Adjust to your actual host/port
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API for personalized city discovery and recommendations.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: WanderWiseAI API
  version: "1.0"
paths:
...
  /poi/city/{cityID}:
    get:
      consumes:
      - application/json
      description: Retrieves all points of interest for a specific city
      parameters:
      - description: City ID
        in: path
        name: cityID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of POIs in the city
          schema:
            items: {}
            type: array
        "400":
          description: Invalid Input
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.Response'
      summary: Get POIs by City ID
      tags:
      - POI
  /poi/favourites:
    delete:
      consumes:
      - application/json
      description: Removes a point of interest from the user's favourites list
      parameters:
      - description: POI ID to remove from favourites
        in: body
        name: poi
        required: true
        schema:
          $ref: '#/definitions/types.AddPoiRequest'
      produces:
      - application/json
      responses:
        "200":
          description: POI removed from favourites successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid Input
          schema:
            $ref: '#/definitions/types.Response'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: Remove POI from Favourites
      tags:
      - POI
    get:
      consumes:
      - application/json
      description: Retrieves all points of interest that the authenticated user has
        marked as favourites
      produces:
      - application/json
      responses:
        "200":
          description: List of favourite POIs
          schema:
            items: {}
            type: array
        "400":
          description: Invalid Input
          schema:
            $ref: '#/definitions/types.Response'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: Get User's Favourite POIs
      tags:
      - POI
    post:
      consumes:
      - application/json
      description: Adds a point of interest to the user's favourites list
      parameters:
      - description: POI ID to add to favourites
        in: body
        name: poi
        required: true
        schema:
          $ref: '#/definitions/types.AddPoiRequest'
      produces:
      - application/json
      responses:
        "201":
          description: POI added to favourites successfully
          schema: {}
        "400":
          description: Invalid Input
          schema:
            $ref: '#/definitions/types.Response'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: Add POI to Favourites
      tags:
      - POI
...
  /user/profile:
    get:
      consumes:
      - application/json
      description: Retrieves the authenticated user's profile information.
      produces:
      - application/json
      responses:
        "200":
          description: User Profile
          schema:
            $ref: '#/definitions/types.UserProfile'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.Response'
        "404":
          description: User Not Found
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: Get User Profile
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Updates the authenticated user's profile information.
      parameters:
      - description: Profile Update Parameters
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/types.UpdateProfileParams'
      produces:
      - application/json
      responses:
        "200":
          description: Profile Updated Successfully
          schema:
            $ref: '#/definitions/types.Response'
        "400":
          description: Invalid Input
          schema:
            $ref: '#/definitions/types.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/types.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.Response'
      security:
      - BearerAuth: []
      summary: Update User Profile
      tags:
      - User
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"